fix/SW-451: disable Request Advisor button after advisor assignment
